--- a/Project.toml
+++ b/Project.toml
@@ -19,7 +19,8 @@ julia = "1"
 
 [extras]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+PooledArrays = "2dfb63ee-cc39-5dd5-95bd-886bf059d720"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Dates", "Test"]
+test = ["Dates", "PooledArrays", "Test"]