This is particularly useful to efficiently convert `PooledArray` columns created by CSV.jl to `CategoricalArray`. `copyto!` is used by `convert` and constructors.

```julia
# Master
julia> using CategoricalArrays, PooledArrays, BenchmarkTools

julia> x = PooledArray(rand(["a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  296.170 μs (10049 allocations: 355.59 KiB)

julia> x = PooledArray(rand([missing, "a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  287.011 μs (8042 allocations: 292.88 KiB)

julia> x = PooledArray(rand(1:100, 10_000));

julia> @btime categorical(x);
  248.418 μs (10281 allocations: 394.07 KiB)

# This PR
julia> x = PooledArray(rand(["a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  34.170 μs (74 allocations: 44.80 KiB)

julia> x = PooledArray(rand([missing, "a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  40.233 μs (74 allocations: 44.83 KiB)

julia> x = PooledArray(rand(1:100, 10_000));

julia> @btime categorical(x);
  59.823 μs (217 allocations: 88.56 KiB)
```

It would be possible to add an even faster path for when the pool is already sorted in the source (and `missing` is in the first position if present), but that's probably not worth the increased complexity unless CSV.jl produces arrays like that. But I don't think it will since for parsing performance it's better to add levels to the pool as they are encountered -- except maybe if there was a detection phase which encountered all levels, allowing to sort them upfront.