@bkamins 2020-06-21
I have left some suggestions, the major is `boundscheck` that I think should not be there.