@nalimilan 2020-06-27
The new version is even faster than before! Though the impact of dropping unused levels is clear (last one):
```julia
julia> x = PooledArray(rand(["a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  21.900 μs (72 allocations: 44.56 KiB)

julia> x = PooledArray(rand([missing, "a", "b", "c", "d"], 10_000));

julia> @btime categorical(x);
  22.804 μs (81 allocations: 45.31 KiB)

julia> x = PooledArray(rand(1:100, 10_000));

julia> @btime categorical(x);
  43.665 μs (208 allocations: 83.13 KiB)

julia> replace!(x, 1=>2);

julia> @btime categorical(x);
  76.009 μs (377 allocations: 114.86 KiB)
```