@nalimilan 2020-06-27
I've realized I wasn't actually testing `SubArray` at all due to `copy(y1)`. :-D Fortunately only a small change was required.

Implementing this optimized method turns out to be much more tricky than I expected. I guess that's the rule with CategoricalArrays...